inline extensions scanner
